% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data
Curate data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
curated dataset
}
\description{
eq_clean_data
Curate data
}
\examples{
\dontrun{
readr::read_delim(raw_data,
                  delim = "\t") \%>\%
                      eq_location_clean()}
}
