% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label
Cleans data}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
a list of labels consisting of Location, Number of Deaths and Magnitude for eache earthquake
}
\description{
eq_create_label
Cleans data
}
\examples{
\dontrun{
readr::read_delim(data,
                  delim = "\t") \%>\%
       eq_location_clean()\%>\%
       dplyr::mutate(popup_text = eq_create_label(.))
                      }
}
