% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean
Cleans data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
cleaned dataset
}
\description{
eq_location_clean
Cleans data
}
\examples{
\dontrun{
readr::read_delim(data,
                  delim = "\t") \%>\%
                      eq_location_clean()}
}
