% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_create_label
Cleans data}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A dataset}

\item{annot_col}{selected column for showing in pop-up}
}
\value{
show earthquake points on maps
}
\description{
eq_create_label
Cleans data
}
\examples{
\dontrun{
readr::read_delim(data,
                  delim = "\t") \%>\%
       eq_clean_data() \%>\%
    dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
    eq_map(annot_col = "DATE")
                      }
}
\references{
https://rstudio.github.io/leaflet/markers.html
}
