% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label
make labels for a timeline of earthquakes for countries}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = F,
  inherit.aes = TRUE,
  ...
)
}
\description{
geom_timeline_label
make labels for a timeline of earthquakes for countries
}
\examples{
\dontrun{
ex_data <- readr::read_delim(raw_data,
                  delim = "\t") \%>\%
                      eq_location_clean() \%>\%
                      filter(YEAR>1500) \%>\%
filter(COUNTRY \%in\% c("IRAN","CHINA","USA","UK"))
ggplot2::ggplot(data = ex_data, aes(x = ex_data$DATE ,
                          y = ex_data$COUNTRY,
                          size = ex_data$EQ_PRIMARY,
                          alpha = ex_data$DEATHS)) +
 geom_timeline_label(n_max = 20)+
 geom_timeline()+
 theme_light()+
 labs(size = "Strength (Richter)",
      alpha = "Number of deaths")+
 xlab("Time") + ylab("")
}
}
