% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{ggproto_timeline}
\alias{ggproto_timeline}
\title{ggproto_timeline
build ggproto object for geom_timeline}
\format{
An object of class \code{ggproto_timeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
ggproto_timeline
}
\description{
ggproto_timeline
build ggproto object for geom_timeline
}
\keyword{datasets}
